+++ src/homematicip/base/enums.py
@@ -351,6 +351,7 @@ class GroupType(AutoNameEnum):
     EXTENDED_LINKED_NOTIFICATION = auto()
     EXTENDED_LINKED_SHUTTER = auto()
     EXTENDED_LINKED_SWITCHING = auto()
+    EXTENDED_LINKED_VENTILATION = auto()
     HEATING = auto()
     HEATING_CHANGEOVER = auto()
     HEATING_COOLING_DEMAND = auto()
@@ -378,6 +379,7 @@ class GroupType(AutoNameEnum):
     SMOKE_ALARM_DETECTION_RULE = auto()
     SWITCHING = auto()
     SWITCHING_PROFILE = auto()
+    VENTILATION_PROFILE = auto()
     DEMAND_CONTROLLED_VENTILATION_GROUP = auto()